use int8 for strand

+++ src/genomicranges/utils.py
@@ -46,7 +46,7 @@ def sanitize_strand_vector(
     if ut.is_list_of_type(strand, str):
         if not set(strand).issubset(["+", "-", "*"]):
             raise ValueError("Values in 'strand' must be either +, - or *.")
-        return np.array([STRAND_MAP[x] for x in strand])
+        return np.array([STRAND_MAP[x] for x in strand], dtype=np.int8)
 
     if ut.is_list_of_type(strand, (int, float, np.int_)):
         if not set(strand).issubset([1, 0, -1]):
